study_buddy_flask_backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
openai
python-dotenv
requests
PyPDF2
